Split dname conversion code to convert module
