+++ src/convert/label.rs
@@ -1,0 +1,33 @@
+use std::io::{Cursor, Read};
+
+use crate::types::dname::label::*;
+
+impl Label {
+    pub fn into_bytes(self) -> Vec<u8> {
+        let size = self.0.len();
+        let mut buf = self.0.into_bytes();
+        buf.splice(0..0, [size as u8]);
+        buf
+    }
+
+    pub fn read_label(bytes: &mut Cursor<&[u8]>, dest: &mut [u8]) -> Result<Self> {
+        bytes.read_exact(dest).map_err(|source| Error::Io {
+            src_amt: bytes.get_ref()[bytes.position() as usize..].len(),
+            dest_amt: dest.len(),
+            source,
+        })?;
+        let label = std::str::from_utf8(dest)
+            .map_err(|source| Error::Convert {
+                bytes: dest.to_vec(),
+                source,
+            })?
+            .to_string();
+        Ok(Self(label))
+    }
+}
+
+impl ToString for Label {
+    fn to_string(&self) -> String {
+        self.0.clone()
+    }
+}
