+++ src/lib.rs
@@ -1,3 +1,4 @@
+mod convert;
 mod types;
 
 pub use types::*;
